pyproject.toml
[project]
name = "test-dbx-delta-share-access"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=3.0.0",
    "python-dotenv>=1.2.1",
]
